+++ src/app/Products/product-list/product-list.component.ts
@@ -1,12 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { CartModule } from 'src/app/Cards/cart.module';
 import { CartService } from 'src/app/Cards/cart.service';
-import { Category } from 'src/app/models/category';
-import { CartModel } from 'src/app/models/cart.model';
 import { ProductModel } from '../../models/product.model';
-import { ProductsService } from '../products.service';
-import { Observable, of } from 'rxjs';
 import { Router } from '@angular/router';
+import { ProductsPromiseService } from '../products-promise.service';
+import { LikedProductsService } from '../products.service';
+import { CartObservableService } from 'src/app/Cards/cart-observable.service';
 
 @Component({
   selector: 'app-product-list',
@@ -17,39 +15,45 @@ export class ProductListComponent implements OnInit {
   products!: Promise<ProductModel[]>;
   name: string = '';
 
-  constructor(private productService: ProductsService,private cartService: CartService,private router: Router) { }
+  constructor(private productService: ProductsPromiseService, private likedService: LikedProductsService, private cartService: CartObservableService, private router: Router) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
   }
 
   onView(product: ProductModel): void {
-    const link = ['/products/product', product.id ];
+    const link = ['/products/product', product.id];
     this.router.navigate(link);
   }
-  onAddToCart(product: ProductModel): void {
-    let cartProduct =  this.cartService.getProducts().find(x => x.name === product.name);
-    if (!cartProduct) {
-      this.cartService.addProduct(product);
-    } else {
-      this.cartService.increaseQuantity(cartProduct,1);
-    }
+  onAddToCart(product: ProductModel, quantity: number = 1): void {
+    let cardProduct = { ...product, quantity };
+    console.log("22")
+    this.cartService.getCartProducts().subscribe(
+      (items) => {
+        console.log(items.includes(cardProduct))
+        if (items.find(x=>x.id === product.id)) {
+           this.cartService.increaseQuantity(cardProduct, 1); 
+          }else { 
+            this.cartService.createCart(cardProduct).subscribe(); 
+          }
+      }
+    );
 
   }
 
   onLikeProduct(product: ProductModel): void {
-    this.productService.likeProduct(product);
-    console.log(this.productService.getLikedProducts());
+    this.likedService.likeProduct(product);
+    console.log(this.likedService.getLikedProducts());
   }
-  
+
   onDislikeProduct(product: ProductModel): void {
-    this.productService.dislikeProduct(product);
-    console.log(this.productService.getLikedProducts());
+    this.likedService.dislikeProduct(product);
+    console.log(this.likedService.getLikedProducts());
   }
 
   onEditProduct(product: ProductModel): void {
-    
-    const link = ['/admin',{ outlets: { admin: ['product','edit',product.id] } }];
+
+    const link = ['/admin', { outlets: { admin: ['product', 'edit', product.id] } }];
     this.router.navigate(link);
   }
 }