+++ src/app/Products/products-promise.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ProductsPromiseService } from './products-promise.service';
+
+describe('ProductsPromiseService', () => {
+  let service: ProductsPromiseService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ProductsPromiseService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});