+++ src/app/admin/components/admin-dashboard/guards/can-resolve-edit.guard.ts
@@ -1,30 +1,28 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve, Router, RouterStateSnapshot } from '@angular/router';
-import { EMPTY } from 'rxjs';
 import { ProductModel } from 'src/app/models/product.model';
-import { ProductsService } from 'src/app/Products/products.service';
+import { ProductsPromiseService } from 'src/app/Products/products-promise.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class CanResolveEditGuard implements Resolve<ProductModel>{
+export class CanResolveEditGuard implements Resolve<ProductModel|undefined>{
 
-  constructor(private router: Router, private productService: ProductsService) { }
-  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<ProductModel> {
-
+  constructor(private router: Router, private productService: ProductsPromiseService) { }
+   async resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<ProductModel|undefined> {
     const id = route.paramMap.get('productID')!;
-    return this.productService.getProductById(id).then(product => {
+     try {
+      const product = await this.productService.getProduct(id);
       if (product) {
         return product;
       } else {
         this.router.navigate(['admin/']);
-        return EMPTY.toPromise();
+        return Promise.resolve(undefined);
       }
-    }
-    ).catch(() => {
+    } catch (e) {
       this.router.navigate(['admin/']);
-      return EMPTY.toPromise();
-    });
+      return Promise.resolve(undefined);
+    }
 
   }
 }