+++ src/app/models/cart.model.ts
@@ -1,4 +1,5 @@
 export interface CartModel {
+  id:number
   name: string;
   description: string;
   price: number;