+++ src/app/models/setting.model.ts
@@ -1,0 +1,3 @@
+export interface SettingModel {
+    id: string;
+}